refactoring pages folder structures
