+++ pages/speakers/index.tsx
@@ -1,8 +1,8 @@
 import { FC, ReactNode } from "react";
 import styled from "styled-components";
-import NavBar from "../components/NavBar";
-import { speakers } from "../data/speakers";
-import { wasmPurple } from "../components/colors";
+import NavBar from "../../components/NavBar";
+import { speakers } from "../../data/speakers";
+import { wasmPurple } from "../../components/colors";
 
 const title = "About the event";
